feat: support extends configuration from oxlint config content (#419)
## Description

This PR introduces an initial implementation to support the extends
property in configuration files and follow the full chain of extended
files.

Currently, it handles merging of only the following fields:
- `rules`
- `plugins`
- `overrides`

These are the only fields currently supported by oxlint, and it doesn’t
appear necessary to support others at this stage.

Fixes #386

---------

Co-authored-by: Sysix <sysix@sysix-coding.de>